Split tpsclient into libpki-tps.so
Most of the code in tpsclient has been moved into libpki-tps.so
(i.e. reverting commit 65d1a36e6dcddf8ab2bf24ad072bbd26753c4a4e)
such that it can be reused by other tools. The main() function
itself has been moved into tpsclient.cpp.
