+++ base/tools/src/main/native/tpsclient/src/main/RA_Conn.cpp
@@ -25,7 +25,7 @@
 #include "plstr.h"
 #include "main/NameValueSet.h"
 #include "main/Util.h"
-#include "RA_Conn.h"
+#include "main/RA_Conn.h"
 #include "apdu/APDU_Response.h"
 #include "apdu/List_Objects_APDU.h"
 #include "apdu/Create_Object_APDU.h"
