+++ base/tools/src/main/native/tpsclient/tools/CMakeLists.txt
@@ -1,0 +1,1 @@
+add_subdirectory(raclient)
