+++ pki.spec
@@ -1729,6 +1729,7 @@ fi
 %{_bindir}/TokenInfo
 %{_datadir}/pki/tools/
 %{_datadir}/pki/lib/p11-kit-trust.so
+%{_libdir}/libpki-tps.so
 %{_mandir}/man1/AtoB.1.gz
 %{_mandir}/man1/AuditVerify.1.gz
 %{_mandir}/man1/BtoA.1.gz
